backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pdfplumber==0.11.4
google-genai==1.0.0
python-dotenv>=1.0.0
python-multipart